+++ textx/__init__.py
@@ -1,7 +1,8 @@
 # flake8: noqa
 from textx.metamodel import metamodel_from_file, metamodel_from_str
 from textx.model import get_children_of_type, get_parent_of_type, \
-    get_model, get_metamodel, get_children, get_location, textx_isinstance
+    get_model, get_metamodel, get_children, get_location, textx_isinstance, \
+    textxerror_wrap
 from textx.exceptions import TextXError, TextXSyntaxError, \
     TextXSemanticError, TextXRegistrationError
 from textx.registration import (LanguageDesc, GeneratorDesc,